+++ prompt_styles/dark-double-line.sh
@@ -23,6 +23,7 @@ fancygit_prompt_builder() {
     branch_end="${bg_none}${none}${bold_none}${s_white}"
     local venv=""
     local path_sign=""
+    local prompt_user=""
 
     # Building prompt
     if [ "$branch_status" != "" ]
@@ -63,7 +64,13 @@ fancygit_prompt_builder() {
         has_unpushed_commits=""
     fi
 
-    prompt_user="${user_at_host}\\u@\\h ${user_at_host_end}"
+    if fg_show_user_at_machine
+    then
+        user_at_host="${white}${bg_dark_gray_01}${bold}"
+        user_at_host_end="${bold_none}${bg_none}${s_darkgray01_bgdarkgray}"
+        prompt_user="${user_at_host}\\u@\\h ${user_at_host_end}"
+    fi
+
     prompt_symbol="\n${user_symbol}\$${user_symbol_end}"
 
     if ! [ -z ${VIRTUAL_ENV} ]