+++ prompt_styles/light-double-line.sh
@@ -12,8 +12,6 @@ fancygit_prompt_builder() {
     . ~/.fancy-git/update_checker.sh && _fancygit_update_checker
 
     # Prompt style
-    user_at_host="${black}${bg_white}${bold}"
-    user_at_host_end="${bold_none}${bg_none}${s_white_bglightgray}"
     user_symbol="${bg_light_gray}${bold}${black}"
     user_symbol_end="${none}${bold_none}${bg_none}${s_lightgray}"
     path="${bg_light_gray}${black}${bold}"
@@ -23,6 +21,7 @@ fancygit_prompt_builder() {
     branch_end="${bg_none}${none}${bold_none}${s_white}"
     local venv=""
     local path_sign=""
+    local prompt_user=""
 
     # Building prompt
     if [ "$branch_status" != "" ]
@@ -63,7 +62,13 @@ fancygit_prompt_builder() {
         has_unpushed_commits=""
     fi
 
-    prompt_user="${user_at_host}\\u@\\h ${user_at_host_end}"
+    if fg_show_user_at_machine
+    then
+        user_at_host="${black}${bg_white}${bold}"
+        user_at_host_end="${bold_none}${bg_none}${s_white_bglightgray}"
+        prompt_user="${user_at_host}\\u@\\h ${user_at_host_end}"
+    fi
+
     prompt_symbol="\n${user_symbol}\$${user_symbol_end}"
 
     if ! [ -z ${VIRTUAL_ENV} ]