+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 > Any trouble, please visit: https://github.com/diogocavilha/fancy-git#troubleshooting-pick
 
+### v6.3.1
+
+- Add one more question to "gd" command.
+- Fix a minor bug when using MacOS. 
+- Add the `fancygit --enable-show-user-at-machine` command for making FancyGit to show the the (user@machine) prompt info.
+- Add the `fancygit --disable-show-user-at-machine` command for making FancyGit not to show the the (user@machine) prompt info.
+- Change the `fancygit --full-path-disable` command to `fancygit --disable-full-path` in orer to keep it organized.
+- Change the `fancygit --full-path-enable` command to `fancygit --enable-full-path` in orer to keep it organized.
+- Add a new font called "Sauce-Code-Pro-Nerd-Font-Complete-Windows-Compatible.ttf" in order to keep compatibility.
+- Update virtualenv status icon.
+- Change stash status icon.
+
 ### v6.3.0
 
 - Update icon for stash.